Implement Tracks Sorting and various
- Renamed TracksListPrefsHelper
- Replaced icons with free-use google icons
- Implemented library "DragTag"

Notes:

The tags still need a lot of polishing. Maybe a darker
color should be used for them and change the icons to white
(and lighter red for ic_remove).

The containers and general OrderByView class need to be
redesigned to look nicer and be more intuitive -- what's
done here is a good first step, but has loads of room
for improvements.